+++ app-demo/src/app/other-page.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+  selector: 'app-other-page',
+  imports: [],
+  template: `<p>other-page works!</p>`,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class OtherPage {}