+++ lib/src/lib/rx-form-submit-options.ts
@@ -4,14 +4,9 @@ import type { Observable } from 'rxjs';
 
 /**
  * Observabled-based version of the Angular `FormSubmitOptions` (which is Promise-based).
- * It can be used:
- * 1. when submitting a form with `rxSubmit()`
- * 2. when using the form `submission` configuration with `rxSubmission()`
- *
- * See the documentation of each property for more details.
+ * It can be used when submitting a form with `rxSubmit()`.
  *
  * @see {@link rxSubmit}
- * @see {@link rxSubmission}
  *
  * @version 21.2.0
  * @experimental
@@ -36,13 +31,15 @@ export interface RxFormSubmitOptions<TRootModel, TSubmittedModel> extends Omit<
       readonly submitted: FieldTree<TSubmittedModel>;
     },
   ) => Observable<TreeValidationResult>;
+
   /**
    * The `DestroyRef` representing the current context.
-   * **It is required when using `rxSubmit()` or `rxSubmission()` outside of an injection context.**
+   * **It is required when using `rxSubmit()` outside of an injection context.**
    * Otherwise, the current `DestroyRef` is injected.
-   * Using `rxSubmit()` or `rxSubmission()` outside an injection context and without providing a `DestroyRef` will throw the `NG0203` runtime error: https://angular.dev/errors/NG0203
+   * Using `rxSubmit()` outside an injection context and without providing a `DestroyRef` will throw the `NG0203` runtime error: https://angular.dev/errors/NG0203
    */
   readonly destroyRef?: Readonly<DestroyRef> | undefined;
+
   // Override Angular options with better signatures for immutability
   readonly onInvalid?: (
     field: FieldTree<TRootModel & TSubmittedModel>,
@@ -51,6 +48,7 @@ export interface RxFormSubmitOptions<TRootModel, TSubmittedModel> extends Omit<
       readonly submitted: FieldTree<TSubmittedModel>;
     },
   ) => void;
+
   readonly ignoreValidators?: Required<
     FormSubmitOptions<TRootModel, TSubmittedModel>
   >['ignoreValidators'];