+++ lib/src/public-api.ts
@@ -3,5 +3,4 @@
  */
 
 export { type RxFormSubmitOptions } from './lib/rx-form-submit-options';
-export { rxSubmission } from './lib/rx-submission';
 export { rxSubmit } from './lib/rx-submit';