+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 21.3.0 (2026-05-14)
+
+- Cancellation now only happens for the `rxSubmit()` observable, not for the `action` observable (see README for more information)
+- As a consequence, `rxSubmission()` has been removed as it no longer provides value
+
 ## 21.2.0 (2026-03-18)
 
 Initial release.